chore: refactor get_public_keys_from_account function to rust client and bump sdk version
